requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.11
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-rest-swagger==2.2.0
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.7
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==75.3.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.4.4
stripe==9.4.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
user-agents==2.2.0
